requirements.txt
asgiref==3.7.2
attrs==23.1.0
beautifulsoup4==4.12.2
blinker==1.6.2
certifi==2023.7.22
cffi==1.16.0
charset-normalizer==3.3.0
click==8.1.7
colorama==0.4.6
contourpy==1.1.1
cycler==0.12.0
Django==4.2.6
django-cors-headers==4.2.0
djangorestframework==3.14.0
EasyProcess==1.1
entrypoint2==1.1
Flask==3.0.0
fonttools==4.43.1
fpdf==1.7.2
h11==0.14.0
idna==3.4
itsdangerous==2.1.2
Jinja2==3.1.2
kiwisolver==1.4.5
MarkupSafe==2.1.3
matplotlib==3.8.0
MouseInfo==0.1.3
mss==9.0.1
mysqlclient==2.2.0
numpy==1.26.0
outcome==1.2.0
packaging==23.2
Pillow==10.0.1
PyAutoGUI==0.9.54
pycparser==2.21
PyGetWindow==0.0.9
PyMsgBox==1.0.9
pyparsing==3.1.1
pyperclip==1.8.2
PyRect==0.2.0
pyscreenshot==3.1
PyScreeze==0.1.29
PySocks==1.7.1
python-dateutil==2.8.2
pytweening==1.0.7
pytz==2023.3.post1
pywhatkit==5.4
requests==2.31.0
six==1.16.0
sniffio==1.3.0
sortedcontainers==2.4.0
soupsieve==2.5
sqlparse==0.4.4
trio==0.22.2
trio-websocket==0.11.1
tzdata==2023.3
urllib3==2.0.6
Werkzeug==3.0.0
wikipedia==1.4.0
wsproto==1.2.0
